update NNUE architecture to 768x16hm -> 4096/2 -> ([16*2 -> 16*2 -> 16*2 -> 1] + [16*2 + 1])x8
--------------------------------------------------
Results of dev vs base (1+0.01, 1t, 16MB, UHO_Lichess_4852_v1.epd):
Elo: 0.76 +/- 2.01, nElo: 1.33 +/- 3.52
LOS: 77.11 %, DrawRatio: 46.63 %, PairsRatio: 1.00
Games: 37386, Wins: 9783, Losses: 9701, Draws: 17902, Points: 18734.0 (50.11 %)
Ptnml(0-2): [339, 4640, 8717, 4594, 403], WL/DD Ratio: 1.01
LLR: 2.89 (100.0%) (-2.25, 2.89) [-3.00, 1.00]
--------------------------------------------------
